+++ core/src/main/java/eu/unicore/ucc/util/Spawner.java
@@ -1,5 +1,7 @@
 package eu.unicore.ucc.util;
 
+import java.util.Properties;
+
 import eu.unicore.ucc.Command;
 import eu.unicore.ucc.UCC;
 
@@ -11,9 +13,10 @@ public class Spawner {
 	private final Command cmd;
 
 	public Spawner(Command base, String[] args) throws Exception {
-		cmd = UCC.initCommand(args, false, base.getProperties());
+		Properties p = base.getProperties();
+		cmd = UCC.initCommand(expandArgs(args,p), false, p);
 		if(cmd!=null){
-			cmd.setProperties(base.getProperties());
+			cmd.setProperties(p);
 			cmd.setPropertiesFile(base.getPropertiesFile());
 		}
 	}
@@ -33,5 +36,26 @@ public class Spawner {
 				UCC.console.setPrefix(pfx);
 			}
 		}
+	}
+
+	private String[] expandArgs(String[] args, Properties p) {
+		String[]res = new String[args.length];
+		for(int i=0; i<args.length; i++) {
+			res[i] = expandVariables(args[i],p);
+		}
+		return res;
+	}
+
+	private String expandVariables(String var, Properties p){
+		for(String key: p.stringPropertyNames()) {
+			if(!var.contains(key))continue;
+			if(var.contains("${"+key+"}")){
+				var = var.replace("${"+key+"}", String.valueOf(p.get(key)));
+			}
+			else if(var.contains("$"+key)){
+				var = var.replace("$"+key, String.valueOf(p.get(key)));				
+			}
+		}
+		return var;
 	}
 }